requirements.txt
# google-adk>=1.20.0
# google-genai>=1.53.0
# googlemaps>=4.10.0
# pydantic>=2.10.0
# python-dotenv>=1.0.0

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-adk>=1.20.0
ag-ui-adk>=0.4.1
google-cloud-aiplatform>=1.38.0
google-genai>=0.3.0
python-dotenv>=1.0.0
# Add any other specific libraries your agent uses (e.g. pandas, numpy)
pandas>=2.0.0
googlemaps>=4.10.0
pydantic>=2.0.0
google-cloud-aiplatform>=1.38.0